pyproject.toml
[project]
name = "agent-for-airflow"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "pydantic-settings>=2.0.0",
    "slack-bolt>=1.18.0",
    "aiohttp>=3.9.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
claude = ["anthropic>=0.40.0"]
openai = ["openai>=1.30.0"]
gemini = ["google-generativeai>=0.8.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "respx>=0.21.0",
]

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
